Reorganize sources
