+++ driver/CMakeLists.txt
@@ -34,62 +34,62 @@ endif ()
 
 message (STATUS "${libname}: UNICODE=${UNICODE} SQL_WCHART_CONVERT=${SQL_WCHART_CONVERT} ODBC_WCHAR=${ODBC_WCHAR} ODBC_CHAR16=${ODBC_CHAR16}")
 
-configure_file (${CMAKE_CURRENT_SOURCE_DIR}/config_cmake.h.in ${CMAKE_CURRENT_BINARY_DIR}/config_cmake${_config_add}.h)
+configure_file (${CMAKE_CURRENT_SOURCE_DIR}/platform/config_cmake.h.in ${CMAKE_CURRENT_BINARY_DIR}/platform/config_cmake${_config_add}.h)
 
 # In order to enable testing, put every non-public symbol to a static library (which is then used by shared library and unit-test binary).
 add_library(${libname}-impl STATIC
     attributes.cpp
-    config.cpp
+    config/config.cpp
     connection.cpp
     descriptor.cpp
     diagnostics.cpp
     driver.cpp
     environment.cpp
     object.cpp
-    read_helpers.cpp
+    utils/read_helpers.cpp
     result_set.cpp
     statement.cpp
     type_info.cpp
-    type_parser.cpp
+    utils/type_parser.cpp
 
     attributes.h
-    config.h
+    config/config.h
     connection.h
     descriptor.h
     diagnostics.h
     driver.h
     environment.h
-    ini_defines.h
-    iostream_debug_helpers.h
+    config/ini_defines.h
+    utils/iostream_debug_helpers.h
     object.h
-    platform.h
-    read_helpers.h
+    platform/platform.h
+    utils/read_helpers.h
     result_set.h
-    scope_guard.h
+    utils/scope_guard.h
     statement.h
-    string_ref.h
+    utils/string_ref.h
     type_info.h
-    type_parser.h
-    unicode_t.h
-    utils.h
+    utils/type_parser.h
+    platform/unicode_t.h
+    utils/utils.h
 )
 
 set (WIN_SOURCES)
 if (WIN32)
     set (WIN_SOURCES
-        "win/dllmain.cpp"
-        "win/resource.h"
-        "win/setup.cpp"
-        "win/resource.rc"
-        "win/${libname}.def"
+        platform/win/dllmain.cpp
+        platform/win/resource.h
+        platform/win/setup.cpp
+        platform/win/resource.rc
+        platform/win/${libname}.def
     )
 endif()
 
 add_library(${libname} SHARED
-    odbc.cpp
-    handles.cpp
-    attr.cpp
-    info.cpp
+    api/odbc.cpp
+    api/impl/handles.cpp
+    api/impl/attr.cpp
+    api/impl/info.cpp
 
     ${WIN_SOURCES}
 )
@@ -180,5 +180,5 @@ set_target_properties(clickhouse-odbc PROPERTIES OUTPUT_NAME "clickhouseodbc")
 set_target_properties(clickhouse-odbcw PROPERTIES OUTPUT_NAME "clickhouseodbcw")
 
 if(CLICKHOUSE_ODBC_TEST)
-    add_subdirectory(ut)
+    add_subdirectory(test)
 endif()
