+++ driver/api/impl/info.cpp
@@ -1,6 +1,6 @@
 #include <Poco/NumberFormatter.h>
 #include "connection.h"
-#include "utils.h"
+#include "utils/utils.h"
 
 #if (ODBCVER >= 0x0380)
 #    define SQL_DRIVER_AWARE_POOLING_CAPABLE 0x00000001L
