+++ driver/api/odbc.cpp
@@ -1,8 +1,8 @@
 #include "driver.h"
-#include "utils.h"
-#include "scope_guard.h"
-#include "string_ref.h"
-#include "type_parser.h"
+#include "utils/utils.h"
+#include "utils/scope_guard.h"
+#include "utils/string_ref.h"
+#include "utils/type_parser.h"
 #include "attributes.h"
 #include "diagnostics.h"
 #include "environment.h"
