+++ driver/environment.cpp
@@ -5,7 +5,7 @@
 #include <ctime>
 #include <sstream>
 #include <string>
-#include "unicode_t.h"
+#include "platform/unicode_t.h"
 
 #if defined(_unix_)
 #    include <pwd.h>
