+++ driver/escaping/lexer.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "string_view.h"
+#include "utils/string_view.h"
 
 #include <deque>
 #include <vector>
