+++ driver/platform/config_cmake.h.in
