+++ driver/platform/linker_script
