+++ driver/platform/win/resource.h
@@ -27,7 +27,7 @@
 #endif
 
 #if defined(UNICODE)
-#    include "config_cmakew.h"
+#    include "platform/config_cmakew.h"
 #else
-#    include "config_cmake.h"
+#    include "platform/config_cmake.h"
 #endif
