+++ driver/platform/win/resource.rc
