+++ driver/test/client_utils.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "platform.h"
+#include "platform/platform.h"
 #include "type_info.h"
 
 namespace {
